+++ src/parser/inline/annotation.test.ts
@@ -22,7 +22,7 @@ describe('Unit: parser/inline/annotation', () => {
     it('ab', () => {
       assert.deepStrictEqual(inspect(parser('((a))')), [['<sup class="annotation">a</sup>'], '']);
       assert.deepStrictEqual(inspect(parser('((ab))')), [['<sup class="annotation">ab</sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('((a\nb))')), [['<sup class="annotation">a b</sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('((a\nb))')), [['<sup class="annotation">a<span class="newline"> </span>b</sup>'], '']);
       assert.deepStrictEqual(inspect(parser('((a\\\nb))')), [['<sup class="annotation">a<br>b</sup>'], '']);
     });