+++ src/parser/inline/strong.test.ts
@@ -22,7 +22,7 @@ describe('Unit: parser/inline/strong', () => {
     it('ab', () => {
       assert.deepStrictEqual(inspect(parser('**a**')), [['<strong>a</strong>'], '']);
       assert.deepStrictEqual(inspect(parser('**ab**')), [['<strong>ab</strong>'], '']);
-      assert.deepStrictEqual(inspect(parser('**a\nb**')), [['<strong>a b</strong>'], '']);
+      assert.deepStrictEqual(inspect(parser('**a\nb**')), [['<strong>a<span class="newline"> </span>b</strong>'], '']);
       assert.deepStrictEqual(inspect(parser('**a\\\nb**')), [['<strong>a<br>b</strong>'], '']);
     });