+++ src/parser/block/util/index.test.ts
@@ -17,8 +17,9 @@ describe('Unit: parser/block/util/indexer', () => {
       assert.deepStrictEqual(inspect(parser(' [#a]]')), undefined);
     });
 
-    it('ab', () => {
+    it('valid', () => {
       assert.deepStrictEqual(inspect(parser(' [#a]')), [['<a href="#index:a" rel="noopener" class="index">a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser(' [#a] ')), [['<a href="#index:a" rel="noopener" class="index">a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('  [#a]')), [['<a href="#index:a" rel="noopener" class="index">a</a>'], '']);
     });